driver: kscan: npcx: add driver support for kscan
* Add low-level keyboard scan driver support.
* Add kscan matrix scan driver support.
