+++ drivers/kscan/CMakeLists.txt
@@ -10,6 +10,8 @@ zephyr_library_sources_ifdef(CONFIG_KSCAN_SDL		kscan_sdl.c)
 zephyr_library_sources_ifdef(CONFIG_KSCAN_HT16K33	kscan_ht16k33.c)
 zephyr_library_sources_ifdef(CONFIG_KSCAN_CST816S	kscan_cst816s.c)
 zephyr_library_sources_ifdef(CONFIG_KSCAN_CAP1203	kscan_cap1203.c)
-zephyr_library_sources_ifdef(CONFIG_KSCAN_NPCX		kscan_npcx.c)
+zephyr_library_sources_ifdef(CONFIG_KSCAN_MATRIX_NPCX	kscan_matrix_npcx.c)
+zephyr_library_sources_ifdef(CONFIG_KSCAN_MATRIX_DRIVER	kscan_matrix.c)
 
 zephyr_library_sources_ifdef(CONFIG_USERSPACE	    kscan_handlers.c)
+zephyr_library_sources_ifdef(CONFIG_USERSPACE	    kscan_matrix_handlers.c)
